Ajuste de errores en dto y entity, el ddl-auto se deja en update

+++ src/main/java/com/estudiantes/entity/Ejercicio.java
@@ -30,7 +30,7 @@ public class Ejercicio implements Serializable {
 
     private boolean estado;
 
-    private Long idProblema;
+    private int idProblema;
 
     @JsonIgnoreProperties("ejercicio")
     @OneToMany(mappedBy = "ejercicio", cascade = CascadeType.ALL, orphanRemoval = true)
